requirements.txt
langgraph
langchain
langchain-core
langchain-ollama
langchain-google-genai
langchain-community
langchain-chroma
langgraph-checkpoint-redis
redis
asyncio
# Tools
ddgs==9.11.1
yfinance==0.2.54
pydantic==2.10.6
loguru
python-dotenv
chromadb
streamlit
plotly
pypdf
rank_bm25
requests
pyyaml
tomli;python_version<"3.11"
# Optional fallback (only needed if embeddings.default = "huggingface")
# langchain-huggingface==1.2.1
# sentence-transformers
